+++ stathat/cmd/alert.go
@@ -5,8 +5,14 @@
 package cmd
 
 import (
-	"github.com/spf13/cobra"
+	"fmt"
+	"net/url"
+
+	"github.com/stathat/cmd/stathat/db"
 	"github.com/stathat/cmd/stathat/intr"
+	"github.com/stathat/cmd/stathat/net"
+
+	"github.com/spf13/cobra"
 )
 
 var alertCmd = &cobra.Command{
@@ -22,20 +28,29 @@ var alertValueCmd = &cobra.Command{
 
 var alertDeltaCmd = &cobra.Command{
 	Use:   "delta",
-	Short: "Create a value alert",
+	Short: "Create a delta alert",
 	RunE:  runAlertDelta,
 }
 
 var alertDataCmd = &cobra.Command{
 	Use:   "data",
-	Short: "Create a value alert",
+	Short: "Create a data alert",
 	RunE:  runAlertData,
 }
 
+var alertCreateFlags struct {
+	TimeWindow string
+	Percentage float64
+	Operator   string
+	TimeDelta  string
+	Threshold  float64
+}
+
 var alertDeleteCmd = &cobra.Command{
-	Use:   "delete",
-	Short: "Delete an alert",
-	RunE:  runAlertDelete,
+	Use:    "delete",
+	Short:  "Delete an alert",
+	RunE:   runAlertDelete,
+	Hidden: true,
 }
 
 var alertListCmd = &cobra.Command{
@@ -45,35 +60,112 @@ var alertListCmd = &cobra.Command{
 }
 
 var alertInfoCmd = &cobra.Command{
-	Use:   "info",
-	Short: "Get info about an alert",
-	RunE:  runAlertInfo,
+	Use:    "info",
+	Short:  "Get info about an alert",
+	RunE:   runAlertInfo,
+	Hidden: true,
 }
 
 func init() {
 	RootCmd.AddCommand(alertCmd)
-	alertCmd.AddCommand(alertValueCmd)
-	alertCmd.AddCommand(alertDeltaCmd)
-	alertCmd.AddCommand(alertDataCmd)
 	alertCmd.AddCommand(alertDeleteCmd)
+	alertCmd.AddCommand(alertInfoCmd)
+
 	alertCmd.AddCommand(alertListCmd)
 	alertListCmd.Flags().BoolVar(&listJSON, "json", false, "display output as JSON")
 	alertListCmd.Flags().BoolVar(&listCSV, "csv", false, "display output as CSV")
-	alertCmd.AddCommand(alertInfoCmd)
+
+	alertCmd.AddCommand(alertDataCmd)
+	addAlertCommonFlags(alertDataCmd)
+
+	alertCmd.AddCommand(alertDeltaCmd)
+	addAlertCommonFlags(alertDeltaCmd)
+	alertDeltaCmd.Flags().Float64Var(&alertCreateFlags.Percentage, "percentage", 50.0, "change percentage")
+	alertDeltaCmd.Flags().StringVar(&alertCreateFlags.Operator, "operator", "gt", "alert operator [gt, lt, dt]")
+	alertDeltaCmd.Flags().StringVar(&alertCreateFlags.TimeDelta, "time_delta", "1d", "compare to time difference")
+
+	alertCmd.AddCommand(alertValueCmd)
+	addAlertCommonFlags(alertValueCmd)
 }
 
 func runAlertValue(cmd *cobra.Command, args []string) error {
-	return nil
+	if len(args) != 1 {
+		return cmd.Usage()
+	}
+	store, err := db.New()
+	if err != nil {
+		return err
+	}
+	stat, ok := store.Lookup(args[0])
+	if !ok {
+		return fmt.Errorf("no stat found for %q", args[0])
+	}
+
+	p := "alerts"
+	v := url.Values{}
+	v.Set("kind", "value")
+	v.Set("stat_id", stat.ID)
+	v.Set("time_window", alertCreateFlags.TimeWindow)
+	v.Set("threshold", ftoa(alertCreateFlags.Threshold))
+	v.Set("operator", alertCreateFlags.Operator)
+	var x interface{}
+	err = net.DefaultAPI.Post(p, v, &x)
+	return err
 }
+
 func runAlertDelta(cmd *cobra.Command, args []string) error {
-	return nil
+	if len(args) != 1 {
+		return cmd.Usage()
+	}
+	store, err := db.New()
+	if err != nil {
+		return err
+	}
+	stat, ok := store.Lookup(args[0])
+	if !ok {
+		return fmt.Errorf("no stat found for %q", args[0])
+	}
+
+	p := "alerts"
+	v := url.Values{}
+	v.Set("kind", "delta")
+	v.Set("stat_id", stat.ID)
+	v.Set("time_window", alertCreateFlags.TimeWindow)
+	v.Set("operator", alertCreateFlags.Operator)
+	v.Set("time_delta", alertCreateFlags.TimeDelta)
+	v.Set("percentage", ftoa(alertCreateFlags.Percentage))
+	var x interface{}
+	err = net.DefaultAPI.Post(p, v, &x)
+	return err
 }
+
 func runAlertData(cmd *cobra.Command, args []string) error {
-	return nil
+	if len(args) != 1 {
+		return cmd.Usage()
+	}
+	store, err := db.New()
+	if err != nil {
+		return err
+	}
+	stat, ok := store.Lookup(args[0])
+	if !ok {
+		return fmt.Errorf("no stat found for %q", args[0])
+	}
+
+	p := "alerts"
+	v := url.Values{}
+	v.Set("kind", "data")
+	v.Set("stat_id", stat.ID)
+	v.Set("time_window", alertCreateFlags.TimeWindow)
+	var x interface{}
+	err = net.DefaultAPI.Post(p, v, &x)
+	return err
 }
+
 func runAlertDelete(cmd *cobra.Command, args []string) error {
 	return nil
 }
+
 func runAlertList(cmd *cobra.Command, args []string) error {
 	if len(args) != 0 {
 		return cmd.Usage()
@@ -104,4 +196,8 @@ func outputAlerts(alerts []intr.Alert) error {
 	}
 
 	return Output(t, enc)
+}
+
+func addAlertCommonFlags(c *cobra.Command) {
+	c.Flags().StringVar(&alertCreateFlags.TimeWindow, "time_window", "1d", "time window")
 }