+++ stathat/CHANGELOG.md
@@ -1,9 +1,16 @@
 # Change Log
 
-## In progress
+## v0.1.1
 
 ### Added
 - `comment` command to add text comments to stats
+- `alert list` command to list all alerts in your account
+- `alert data`
+- `alert delta`
+- `alert value`
+
+### Changed
+- `stathat summary` output tweaks to include timeframe, stat kind
 
 ## v0.1.0 (soft public repo available)